feat: app.js 생성
